Refactor: Split GoMath code into smaller util files
