+++ src/Goban/SVGRenderer.ts
@@ -21,14 +21,20 @@ import { AdHocFormat } from "engine/formats/AdHocFormat";
 //import { GobanCore, GobanSelectedThemes, GobanMetrics, GOBAN_FONT } from "./GobanCore";
 import { GobanConfig } from "../GobanBase";
 import { GoEngine } from "engine";
-import * as GoMath from "engine/GoMath";
 import { MoveTree } from "engine/MoveTree";
 import { GobanTheme, THEMES } from "./themes";
 import { MoveTreePenMarks } from "engine/MoveTree";
 import { getRelativeEventPosition } from "./canvas_utils";
 import { _ } from "engine/translate";
 import { formatMessage, MessageID } from "engine/messages";
-import { color_blend, getRandomInt } from "engine/util";
+import {
+    color_blend,
+    encodeMove,
+    encodeMoves,
+    encodePrettyXCoordinate,
+    getRandomInt,
+    makeStringMatrix,
+} from "engine/util";
 import { callbacks } from "./callbacks";
 import { Goban, GobanMetrics, GobanSelectedThemes } from "./Goban";
 
@@ -802,7 +808,7 @@ export class SVGRenderer extends Goban implements GobanSVGInterface {
             }
             const sent = this.sendMove({
                 game_id: this.game_id,
-                move: GoMath.encodeMove(x, y),
+                move: encodeMove(x, y),
             });
             if (sent) {
                 this.playMovementSound();
@@ -838,7 +844,7 @@ export class SVGRenderer extends Goban implements GobanSVGInterface {
                     this.socket.send("game/removed_stones/set", {
                         game_id: this.game_id,
                         removed: removed,
-                        stones: GoMath.encodeMoves(group),
+                        stones: encodeMoves(group),
                     });
                 }
             } else if (this.mode === "puzzle") {
@@ -2799,7 +2805,7 @@ export class SVGRenderer extends Goban implements GobanSVGInterface {
                                     this.square_size +
                                 this.square_size / 2;
                             const y = j * this.square_size + this.square_size / 2;
-                            place(GoMath.encodePrettyXCoordinate(c), x, y);
+                            place(encodePrettyXCoordinate(c), x, y);
                         }
                         break;
                     case "1-1":
@@ -2980,7 +2986,7 @@ export class SVGRenderer extends Goban implements GobanSVGInterface {
             this.__draw_state.length !== this.height ||
             this.__draw_state[0].length !== this.width
         ) {
-            this.__draw_state = GoMath.makeStringMatrix(this.width, this.height);
+            this.__draw_state = makeStringMatrix(this.width, this.height);
         }
 
         for (let j = this.bounds.top; j <= this.bounds.bottom; ++j) {
