+++ src/engine/util/move_encoding.ts
@@ -14,104 +14,13 @@
  * limitations under the License.
  */
 
-import { JGOFIntersection, JGOFMove, JGOFNumericPlayerColor } from "./formats/JGOF";
-import { AdHocPackedMove } from "./formats/AdHocFormat";
-
-export type Intersection = JGOFIntersection;
-export type Matrix<T> = T[][];
-export type NumberMatrix = Matrix<number>;
-export type StringMatrix = Matrix<string>;
-
-export function makeMatrix<T = number>(width: number, height: number, initialValue: T): Matrix<T> {
-    const ret: Matrix<T> = [];
-    for (let y = 0; y < height; ++y) {
-        ret.push([]);
-        for (let x = 0; x < width; ++x) {
-            ret[y].push(initialValue);
-        }
-    }
-    return ret;
-}
-export function makeStringMatrix(
-    width: number,
-    height: number,
-    initialValue: string = "",
-): StringMatrix {
-    const ret: StringMatrix = [];
-    for (let y = 0; y < height; ++y) {
-        ret.push([]);
-        for (let x = 0; x < width; ++x) {
-            ret[y].push(initialValue);
-        }
-    }
-    return ret;
-}
-export function makeObjectMatrix<T>(width: number, height: number): Array<Array<T>> {
-    const ret = new Array<Array<T>>(height);
-    for (let y = 0; y < height; ++y) {
-        const row = new Array<T>(width);
-        for (let x = 0; x < width; ++x) {
-            row[x] = {} as T;
-        }
-        ret[y] = row;
-    }
-    return ret;
-}
-export function makeEmptyObjectMatrix<T>(width: number, height: number): Array<Array<T>> {
-    const ret = new Array<Array<T>>(height);
-    for (let y = 0; y < height; ++y) {
-        const row = new Array<T>(width);
-        ret[y] = row;
-    }
-    return ret;
-}
-
-/* Lower case, includes i, used for our string encoding of moves */
-const COORDINATE_SEQUENCE = "abcdefghijklmnopqrstuvwxyz";
-
-/** Decodes a single coordinate to a number */
-export function decodeCoordinate(ch: string): number {
-    return COORDINATE_SEQUENCE.indexOf(ch?.toLowerCase());
-}
-
-/** Encodes a single coordinate to a number */
-export function encodeCoordinate(coor: number): string {
-    return COORDINATE_SEQUENCE[coor];
-}
-
-/* Upper case, and doesn't have I */
-const PRETTY_COORDINATE_SEQUENCE = "ABCDEFGHJKLMNOPQRSTUVWXYZ";
-
-/** Decodes the pretty X coordinate to a number */
-export function decodePrettyXCoordinate(ch: string): number {
-    return PRETTY_COORDINATE_SEQUENCE.indexOf(ch?.toUpperCase());
-}
-
-/** Encodes an X coordinate to a display encoding */
-export function encodePrettyXCoordinate(coor: number): string {
-    return PRETTY_COORDINATE_SEQUENCE[coor];
-}
-
-export function prettyCoordinates(x: number, y: number, board_height: number): string {
-    if (x >= 0) {
-        return encodePrettyXCoordinate(x) + ("" + (board_height - y));
-    }
-    return "pass";
-}
-export function decodeGTPCoordinates(move: string, width: number, height: number): JGOFMove {
-    if (move === ".." || move.toLowerCase() === "pass") {
-        return { x: -1, y: -1 };
-    }
-    let y = height - parseInt(move.substr(1));
-    const x = decodePrettyXCoordinate(move[0]);
-    if (x === -1) {
-        y = -1;
-    }
-    return { x, y };
-}
-export function decodePrettyCoordinates(move: string, height: number): JGOFMove {
-    return decodeGTPCoordinates(move, -1, height);
-}
+import { JGOFMove, JGOFNumericPlayerColor, AdHocPackedMove } from "../formats";
+import {
+    decodeCoordinate,
+    decodeGTPCoordinates,
+    decodePrettyXCoordinate,
+    encodeCoordinate,
+} from "./coordinates";
 
 /**
  *  Decodes any of the various ways we express moves that we've accumulated over the years into
@@ -331,54 +240,6 @@ export function encodeMovesToArray(moves: Array<JGOFMove>): Array<AdHocPackedMov
     return ret;
 }
 
-/**
- * Removes superfluous fields from the JGOFMove objects, such as
- * edited=false and color=0. This does not modify the original array.
- */
-export function trimJGOFMoves(arr: Array<JGOFMove>): Array<JGOFMove> {
-    return arr.map((o) => {
-        const r: JGOFMove = {
-            x: o.x,
-            y: o.y,
-        };
-        if (o.edited) {
-            r.edited = o.edited;
-        }
-        if (o.color) {
-            r.color = o.color;
-        }
-        if (o.timedelta) {
-            r.timedelta = o.timedelta;
-        }
-        return r;
-    });
-}
-
-/** Returns a sorted move string, this is used in our stone removal logic */
-export function sortMoves(moves: string, width: number, height: number): string;
-export function sortMoves(moves: JGOFMove[], width: number, height: number): JGOFMove[];
-export function sortMoves(
-    moves: string | JGOFMove[],
-    width: number,
-    height: number,
-): string | JGOFMove[] {
-    if (moves instanceof Array) {
-        return moves.sort((a, b) => {
-            const av = (a.edited ? 1 : 0) * 10000 + a.x + a.y * 100;
-            const bv = (b.edited ? 1 : 0) * 10000 + b.x + b.y * 100;
-            return av - bv;
-        });
-    } else {
-        const arr = decodeMoves(moves, width, height);
-        arr.sort((a, b) => {
-            const av = (a.edited ? 1 : 0) * 10000 + a.x + a.y * 100;
-            const bv = (b.edited ? 1 : 0) * 10000 + b.x + b.y * 100;
-            return av - bv;
-        });
-        return encodeMoves(arr);
-    }
-}
-
 // OJE Sequence format is '.root.K10.Q1'  ...
 export function ojeSequenceToMoves(sequence: string): Array<JGOFMove> {
     const plays = sequence.split(".");
@@ -395,65 +256,4 @@ export function ojeSequenceToMoves(sequence: string): Array<JGOFMove> {
     });
 
     return moves;
-}
-
-// This is intended to be an "easy to understand" method of generating a unique id
-// for a board position.
-
-// The "id" is the list of all the positions of the stones, black first then white,
-// separated by a colon.
-
-// There are in fact 8 possible ways to list the positions (all the rotations and
-// reflections of the position).   The id is the lowest (alpha-numerically) of these.
-
-// Colour independence for the position is achieved by takeing the lexically lower
-// of the ids of the position with black and white reversed.
-
-// The "easy to understand" part is that the id can be compared visually to the
-// board position
-
-// The downside is that the id string can be moderately long for boards with lots of stones
-
-type BoardTransform = (x: number, y: number) => { x: number; y: number };
-
-export function positionId(
-    position: Array<Array<JGOFNumericPlayerColor>>,
-    height: number,
-    width: number,
-): string {
-    // The basic algorithm is to list where each of the stones are, in a long string.
-    // We do this once for each transform, selecting the lowest (lexically) as we go.
-
-    const transforms: Array<BoardTransform> = [
-        (x, y) => ({ x, y }),
-        (x, y) => ({ x, y: height - y - 1 }),
-        (x, y) => ({ x: y, y: x }),
-        (x, y) => ({ x: y, y: width - x - 1 }),
-        (x, y) => ({ x: height - y - 1, y: x }),
-        (x, y) => ({ x: height - y - 1, y: width - x - 1 }),
-        (x, y) => ({ x: width - x - 1, y }),
-        (x, y) => ({ x: width - x - 1, y: height - y - 1 }),
-    ];
-
-    const ids = [];
-
-    for (const transform of transforms) {
-        let black_state = "";
-        let white_state = "";
-        for (let x = 0; x < width; x++) {
-            for (let y = 0; y < height; y++) {
-                const c = transform(x, y);
-                if (position[x][y] === JGOFNumericPlayerColor.BLACK) {
-                    black_state += encodeMove(c.x, c.y);
-                }
-                if (position[x][y] === JGOFNumericPlayerColor.WHITE) {
-                    white_state += encodeMove(c.x, c.y);
-                }
-            }
-        }
-
-        ids.push(`${black_state}.${white_state}`);
-    }
-
-    return ids.reduce((prev, current) => (current < prev ? current : prev));
 }
