+++ tsconfig.json
@@ -33,7 +33,12 @@
         "sourceMap": true,
         "jsx": "react"
     },
-    "files": ["src/index.ts", "src/engine/index.ts", "jest.config.ts"],
+    "files": [
+        "src/index.ts",
+        "src/engine/index.ts",
+        "jest.config.ts",
+        "./src/engine/util/getRandomInt.ts"
+    ],
     "include": ["test/**/*.ts"],
     "ts-node": {
         "require": ["tsconfig-paths/register"]
